instance/requirements.txt
Flask
Flask-SQLAlchemy
easyocr
pandas
numpy
Pillow
pdfplumber
